# docs/public-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Threa Public API",
    "version": "1.0.0",
...
  },
  "servers": [{ "url": "https://app.threa.io", "description": "Production" }],
  "security": [{ "apiKey": [] }],
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/me/bots": {
      "get": {
        "operationId": "listMyBots",
        "summary": "List my personal bots",
        "tags": ["Identity"],
        "description": "For user-scoped keys: lists the authenticated user's personal bots, optionally filtered by trait. Used by the frontend to enumerate quick-switcher commands. Bot-scoped keys receive 403.",
        "security": [{ "apiKey": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Workspace ID (prefixed ULID)"
          },
          {
            "name": "traits",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["interactive"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "workspaceId": { "type": "string" },
                          "traits": { "type": "array", "items": { "type": "string", "enum": ["interactive"] } },
                          "slug": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                          "name": { "type": "string" },
                          "description": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                          "avatarEmoji": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                          "avatarUrl": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                          "archivedAt": { "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }] },
                          "createdAt": { "type": "string", "format": "date-time" },
                          "updatedAt": { "type": "string", "format": "date-time" },
                          "type": { "type": "string", "const": "personal" },
                          "ownerUserId": { "type": "string" }
                        },
                        "required": [
                          "id",
                          "workspaceId",
                          "traits",
                          "slug",
                          "name",
                          "description",
                          "avatarEmoji",
                          "avatarUrl",
                          "archivedAt",
                          "createdAt",
                          "updatedAt",
                          "type",
                          "ownerUserId"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["data"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "details": {
                      "type": "object",
                      "propertyNames": { "type": "string" },
                      "additionalProperties": { "type": "array", "items": { "type": "string" } }
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": { "description": "Missing or invalid API key" },
          "403": { "description": "Insufficient permissions or stream not accessible" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key created by a workspace admin in Settings > API Keys. Prefix: `thr_`."
      }
    }
  },
...
}